Moved all components under component directory
