+++ packages/react-mutation-mapper/src/component/hotspot/HotspotInfo.tsx
