+++ packages/react-mutation-mapper/src/component/hotspot/hotspotInfo.module.scss
