+++ packages/react-mutation-mapper/src/component/mutationMapper/MutationMapper.tsx
@@ -4,25 +4,25 @@ import * as React from "react";
 import {ReactNode} from "react";
 import {TableProps} from "react-table";
 
-import {DefaultPubMedCache} from "./cache/DefaultPubMedCache";
-import FilterResetPanel from "./component/FilterResetPanel";
-import {DataFilter} from "./model/DataFilter";
-import {ApplyFilterFn, FilterApplier} from "./model/FilterApplier";
-import {LollipopPlotControlsConfig} from "./model/LollipopPlotControlsConfig";
-import {MobxCache} from "./model/MobxCache";
-import {Mutation} from "./model/Mutation";
-import MutationMapperStore from "./model/MutationMapperStore";
-import {DefaultLollipopPlotControlsConfig} from "./store/DefaultLollipopPlotControlsConfig";
-import DefaultMutationMapperStore from "./store/DefaultMutationMapperStore";
-import {initDefaultTrackVisibility} from "./util/TrackUtils";
-import {getDefaultWindowInstance} from "./util/DefaultWindowInstance";
-import {ColumnSortDirection, DataTableColumn} from "./DataTable";
+import {DefaultPubMedCache} from "../../cache/DefaultPubMedCache";
+import FilterResetPanel from "./FilterResetPanel";
+import {DataFilter} from "../../model/DataFilter";
+import {ApplyFilterFn, FilterApplier} from "../../model/FilterApplier";
+import {LollipopPlotControlsConfig} from "../../model/LollipopPlotControlsConfig";
+import {MobxCache} from "../../model/MobxCache";
+import {Mutation} from "../../model/Mutation";
+import MutationMapperStore from "../../model/MutationMapperStore";
+import {DefaultLollipopPlotControlsConfig} from "../../store/DefaultLollipopPlotControlsConfig";
+import DefaultMutationMapperStore from "../../store/DefaultMutationMapperStore";
+import {initDefaultTrackVisibility} from "../../util/TrackUtils";
+import {getDefaultWindowInstance} from "../../util/DefaultWindowInstance";
+import {ColumnSortDirection, DataTableColumn} from "../dataTable/DataTable";
 import DefaultMutationRateSummary, {MutationRate} from "./DefaultMutationRateSummary";
-import DefaultMutationTable from "./DefaultMutationTable";
+import DefaultMutationTable from "../mutationTable/DefaultMutationTable";
 import GeneSummary from "./GeneSummary";
-import LollipopMutationPlot from "./LollipopMutationPlot";
-import {DEFAULT_MUTATION_COLUMNS} from "./MutationColumnHelper";
-import {TrackDataStatus, TrackName, TrackVisibility} from "./TrackSelector";
+import LollipopMutationPlot from "../lollipopMutationPlot/LollipopMutationPlot";
+import {DEFAULT_MUTATION_COLUMNS} from "../mutationTable/MutationColumnHelper";
+import {TrackDataStatus, TrackName, TrackVisibility} from "../track/TrackSelector";
 
 export type MutationMapperProps = {
     hugoSymbol?: string;
