+++ packages/react-mutation-mapper/src/component/mutationTable/DefaultMutationTable.tsx
@@ -5,20 +5,20 @@ import {observer} from "mobx-react";
 import * as React from "react";
 import {Column} from "react-table";
 
-import Annotation, {getAnnotationData} from "./component/column/Annotation";
-import ClinVar from "./component/column/ClinVar";
-import Gnomad, {getMyVariantInfoData} from "./component/column/Gnomad";
-import {MutationFilterValue} from "./filter/MutationFilter";
-import {IHotspotIndex} from "./model/CancerHotspot";
-import {DataFilterType} from "./model/DataFilter";
-import {MobxCache} from "./model/MobxCache";
-import {Mutation} from "./model/Mutation";
-import {CancerGene, IOncoKbData} from "./model/OncoKb";
-import {RemoteData} from "./model/RemoteData";
-import {SimpleCache} from "./model/SimpleCache";
-import {findNonTextInputFilters, TEXT_INPUT_FILTER_ID} from "./util/FilterUtils";
-import {getRemoteDataGroupStatus} from "./util/RemoteDataUtils";
-import DataTable, {DataTableColumn, DataTableProps} from "./DataTable";
+import Annotation, {getAnnotationData} from "../column/Annotation";
+import ClinVar from "../column/ClinVar";
+import Gnomad, {getMyVariantInfoData} from "../column/Gnomad";
+import {MutationFilterValue} from "../../filter/MutationFilter";
+import {IHotspotIndex} from "../../model/CancerHotspot";
+import {DataFilterType} from "../../model/DataFilter";
+import {MobxCache} from "../../model/MobxCache";
+import {Mutation} from "../../model/Mutation";
+import {CancerGene, IOncoKbData} from "../../model/OncoKb";
+import {RemoteData} from "../../model/RemoteData";
+import {SimpleCache} from "../../model/SimpleCache";
+import {findNonTextInputFilters, TEXT_INPUT_FILTER_ID} from "../../util/FilterUtils";
+import {getRemoteDataGroupStatus} from "../../util/RemoteDataUtils";
+import DataTable, {DataTableColumn, DataTableProps} from "../dataTable/DataTable";
 import {MutationColumn} from "./MutationColumnHelper";
 
 import './defaultMutationTable.scss';
