+++ packages/react-mutation-mapper/src/component/mutationTable/defaultMutationTable.scss
