+++ packages/react-mutation-mapper/src/component/ptm/PtmAnnotationTable.tsx
@@ -5,8 +5,8 @@ import ReactTable, {Column} from "react-table";
 import {computed} from "mobx";
 import {observer} from "mobx-react";
 
-import {Cache, MobxCache} from "./model/MobxCache";
-import {PostTranslationalModification} from "./model/PostTranslationalModification";
+import {Cache, MobxCache} from "../../model/MobxCache";
+import {PostTranslationalModification} from "../../model/PostTranslationalModification";
 import PtmReferenceList from "./PtmReferenceList";
 
 export type PtmSummaryTableProps = {
