+++ packages/react-mutation-mapper/src/component/ptm/PtmReferenceList.tsx
@@ -2,8 +2,8 @@ import {DefaultTooltip} from "cbioportal-frontend-commons";
 import {observer} from "mobx-react";
 import * as React from "react";
 
-import ReferenceList from "./component/oncokb/ReferenceList";
-import {Cache} from "./model/MobxCache";
+import ReferenceList from "../oncokb/ReferenceList";
+import {Cache} from "../../model/MobxCache";
 
 type PtmReferenceListProps = {
     pubmedIds: string[];
